Time: 358 ms (54.60%) | Memory: 175.2 MB (53.37%) - LeetSync

+++ 3618-find-the-original-typed-string-ii/find-the-original-typed-string-ii.cpp
@@ -17,10 +17,10 @@ public:
         }
         n = arr.size();
         if(n >= k) return total;
-        vector<long long> prev(k+1, 1);
+        vector<int> prev(k+1, 1);
         k--;
         for(int i = n-1; i >= 0; i--){
-            vector<long long> curr(k + 1, 0);
+            vector<int> curr(k + 1, 0);
             for(int j = k-1; j >= 0; j--){
                 prev[j] = (prev[j] + prev[j+1]) % mod; 
             }
